Restore original naumann demo
