--- demos/barcoding/actors/icon.qrc
@@ -1,5 +1,0 @@
-<RCC>
-  <qresource prefix="newPrefix">
-    <file>rainbow_dir.png</file>
-  </qresource>
-</RCC>
