--- demos/barcoding/neuro_class.csv
@@ -1,9 +1,0 @@
-neuro_type,RL,LR,LL,xL,Lx,RR,Rx,xR
-oB,0,1,1,1,1,0,0,0
-B,0,0,1,1,1,0,0,0
-iB,1,0,1,1,1,0,0,0
-ioB,1,1,1,1,1,0,0,0
-iMm,1,0,1,1,0,0,0,0
-Mm,0,0,1,1,0,0,0,0
-oMm,0,1,1,0,1,0,0,0
-S,0,0,1,0,0,0,0,0
