+++ demos/naumann/naumann_demo.yaml
@@ -7,14 +7,14 @@ actors:
   Acquirer:
     package: actors.acquire_tbif
     class: TbifAcquirer
-    filename: ./demos/naumann/data/08-17-14_1437_F1_6dpfCOMPLETESET_WB_overclimbing_z-1.tbif
+    filename: data/08-17-14_1437_F1_6dpfCOMPLETESET_WB_overclimbing_z-1.tbif
     framerate: 3.6
 
   Processor:
     package: actors.processor
     class: CaimanProcessor
-    init_filename: demos/naumann/data/tbif_ex_crop.h5
-    config_file: ./demos/naumann/naumann_caiman_params.txt
+    init_filename: data/tbif_ex_crop.h5
+    config_file: naumann_caiman_params.txt
 
   Visual:
     package: actors.visual_model
